added bitmap from linux source

+++ include/string.h
@@ -1,10 +1,14 @@
 #ifndef __STRING_H__
 #define __STRING_H__ 1
 
-int strlen(const char* s);
+#include <env.h>
+
+size_t strlen(const char* s);
 
 char* strcpy(char* dst, const char* src);
-char* strncpy(char* dst, const char* src, int n);
+char* strncpy(char* dst, const char* src, size_t n);
+
+
 
 char* u32_to_str(unsigned int value, char *str, int base);
 char* u64_to_str(unsigned long long value, char *str, int base);
